\name{cirdata}
\alias{cirdata}

\docType{data}
\title{Random Circular data.}

\description{An example data set.   The data have been created randomly for the purpose of running some examples of the functions of the package \bold{isocir}.}

\usage{data(cirdata)}

\details{
  Every data of the vector come from the unit circle \eqn{[0,2\pi]}{[0,2pi]}.  
}

\references{
 
Mardia, K. and Jupp, P. (2000).
  \emph{Directional Statistics}, Chichester: Wiley.

}

\seealso{
\code{\link{cirgenes}}, \code{\link{datareplic}}.
}


\examples{
data(rancir)
}

\keyword{datasets}
\keyword{circular}
