\name{datareplic}
\alias{datareplic}

\docType{data}
\title{Random Circular data.}

\description{An example data set.   The data have been created randomly for the purpose of running some examples of the functions of the package \bold{isocir}.}

\usage{data(datareplic)}

\details{
  Every data of the matrix come from the unit circle \eqn{[0,2\pi]}{[0,2pi]}.  

The data in the matrix \code{datareplic} are random angular measurements.  Its columns are the replications and its rows are the populations.
}

\references{
 
Mardia, K. and Jupp, P. (2000).
  \emph{Directional Statistics}, Chichester: Wiley.

}

\seealso{
\code{\link{cirgenes}}, \code{\link{cirdata}}.
}


\examples{
data(rancir)
}

\keyword{datasets}
\keyword{circular}
